==> 05.Network_analysis.r <==
# https://psbweb05.psb.ugent.be/conet/microbialnetworks/spieceasi.php

library(phyloseq)
library(tidytree)
#install_github("zdk123/SpiecEasi")
library(SpiecEasi)
library(devtools)
install_github("hallucigenia-sparsa/seqtime") 
library(seqtime)
